Release 1.2.5
- Fix recording blocking main thread
- Fix time series chart with correct date

+++ Controller/Adminhtml/Recording/Delete.php
@@ -13,6 +13,7 @@ namespace Mageproxy\Connector\Controller\Adminhtml\Recording;
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
 use Magento\Framework\App\Action\HttpGetActionInterface;
+use Mageproxy\Connector\Api\OptimizationManagerInterface;
 use Mageproxy\Connector\Api\RecordingManagerInterface;
 use Mageproxy\Connector\Api\RecordingRepositoryInterface;
 
@@ -31,18 +32,26 @@ class Delete extends Action implements HttpGetActionInterface
     private RecordingRepositoryInterface $recordingRepository;
 
     /**
+     * @var \Mageproxy\Connector\Api\OptimizationManagerInterface
+     */
+    private OptimizationManagerInterface $optimizationManager;
+
+    /**
      * @param \Magento\Backend\App\Action\Context $context
      * @param \Mageproxy\Connector\Api\RecordingManagerInterface $recordingManager
      * @param \Mageproxy\Connector\Api\RecordingRepositoryInterface $recordingRepository
+     * @param \Mageproxy\Connector\Api\OptimizationManagerInterface $optimizationManager
      */
     public function __construct(
         Context $context,
         RecordingManagerInterface $recordingManager,
-        RecordingRepositoryInterface $recordingRepository
+        RecordingRepositoryInterface $recordingRepository,
+        OptimizationManagerInterface $optimizationManager
     ) {
         parent::__construct($context);
         $this->recordingManager = $recordingManager;
         $this->recordingRepository = $recordingRepository;
+        $this->optimizationManager = $optimizationManager;
     }
 
     public function execute()
@@ -51,6 +60,10 @@ class Delete extends Action implements HttpGetActionInterface
 
         try {
             $recording = $this->recordingRepository->getById($id);
+            if ($this->optimizationManager->deploymentInProgress($recording->getStoreId())) {
+                $this->messageManager->addErrorMessage(__('The recording is deployed and cannot be deleted. First revert the deployment before deleting.'));
+                return $this->_redirect('*/*/');
+            }
             $this->recordingManager->delete($recording);
         } catch (\Exception $e) {
             $this->messageManager->addErrorMessage(__('An error occurred while deleting the recording.'));
